+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.4.0.0
+
+* Add removeTimestamp, removeTimestampC
+* Add setTimestampFromDurations, setTimestampFromDurationsC
+* Merge Data.MediaBus.Conduit.Timing into Data.MediaBus.Conduit.SyncStream
+* Rename setSequenceNumbersAndTimestamps to setSequenceNumberAndTimestamp
+* Rename convertTicksC' to convertTimestampC'
+* Rename mapTicksC to mapTimestampC
+* Rename deriveFrameTimestamp to setTimestampFromDurationsC
+* Improve SyncStream documentation
+* Add conduit wrapper functions for the SyncStream functions
+
 # 0.3.3.0
 
 * Remove `Reframer`